requirements-ppe.txt
ultralytics>=8.4.26
opencv-python>=4.10.0
deep-sort-realtime>=1.3.2
numpy>=1.24.0
torch>=2.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
transformers>=4.57.0
Pillow>=10.0.0
